requirements.txt
ccxt
pandas<3.0.0
numpy
matplotlib
python-dotenv
scikit-learn
joblib
streamlit
